+++ src/main/java/io/zerogone/blog/controller/BlogController.java
@@ -1,7 +1,7 @@
-package io.zerogone.controller;
+package io.zerogone.blog.controller;
 
-import io.zerogone.model.BlogName;
-import io.zerogone.model.dto.BlogDto;
+import io.zerogone.blog.model.BlogName;
+import io.zerogone.blog.model.BlogDto;
 import io.zerogone.service.search.SearchService;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.annotation.Validated;
@@ -13,18 +13,18 @@ import javax.validation.Valid;
 
 @Controller
 @Validated
-public class BlogSettingController {
+public class BlogController {
     private final SearchService<BlogName, BlogDto> searchService;
 
-    public BlogSettingController(SearchService<BlogName, BlogDto> searchService) {
+    public BlogController(SearchService<BlogName, BlogDto> searchService) {
         this.searchService = searchService;
     }
 
-    @GetMapping("{name}/setting")
-    public ModelAndView handleBlogSettingPage(@PathVariable @Valid BlogName name) {
+    @GetMapping("{name}")
+    public ModelAndView handleBlogMainPage(@PathVariable @Valid BlogName name) {
         ModelAndView modelAndView = new ModelAndView();
         modelAndView.addObject("blog", searchService.search(name));
-        modelAndView.setViewName("blog_setting");
+        modelAndView.setViewName("main");
         return modelAndView;
     }
 }
