+++ src/main/java/io/zerogone/blog/controller/BlogCreateController.java
@@ -1,7 +1,7 @@
-package io.zerogone.controller.api;
+package io.zerogone.blog.controller;
 
-import io.zerogone.model.dto.BlogDto;
-import io.zerogone.model.dto.BlogMemberDto;
+import io.zerogone.blog.model.BlogDto;
+import io.zerogone.blog.model.BlogMemberDto;
 import io.zerogone.user.model.UserDto;
 import io.zerogone.model.entity.MemberRole;
 import io.zerogone.service.create.CreateWithImageService;
