+++ src/main/java/io/zerogone/blog/service/BlogIdSearchService.java
@@ -1,9 +1,10 @@
-package io.zerogone.service.search;
+package io.zerogone.blog.service;
 
 import io.zerogone.exception.NotExistDataException;
-import io.zerogone.model.dto.BlogDto;
-import io.zerogone.model.entity.Blog;
-import io.zerogone.repository.BlogDao;
+import io.zerogone.blog.model.BlogDto;
+import io.zerogone.blog.model.Blog;
+import io.zerogone.blog.BlogDao;
+import io.zerogone.service.search.SearchService;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.stereotype.Service;
