+++ src/main/java/io/zerogone/controller/api/PostCreateController.java
@@ -1,7 +1,7 @@
 package io.zerogone.controller.api;
 
 import io.zerogone.exception.NotAuthorizedException;
-import io.zerogone.model.dto.BlogDto;
+import io.zerogone.blog.model.BlogDto;
 import io.zerogone.model.dto.PostDto;
 import io.zerogone.user.model.UserDto;
 import io.zerogone.service.create.CreateService;
