+++ src/main/java/io/zerogone/user/model/UserDto.java
@@ -1,6 +1,6 @@
 package io.zerogone.user.model;
 
-import io.zerogone.model.dto.BlogDto;
+import io.zerogone.blog.model.BlogDto;
 import io.zerogone.validator.Login;
 import io.zerogone.validator.NewEntity;
 import org.hibernate.validator.constraints.Range;
